(#3) Extracted cookies to CookiesUtils.jsx
