+++ time_cost/package.json
@@ -14,8 +14,10 @@
   },
   "dependencies": {
     "bootstrap": "^5.3.0",
+    "crypto-js": "^4.1.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
+    "react-cookie": "^4.1.1",
     "react-currency-input-field": "^3.6.10",
     "react-dom": "^18.2.0"
   },
